refactor: rename V1-prefixed modules to match their post-V2 reality
Three modules carried '_incremental_*' prefixes that no longer
described what they do post-B6 (when V1's incremental machinery was
deleted). Renamed to honest names:

  _incremental_orchestrator.py → DELETED
    Was already a thin delegate. Inlined its 4-line body into
    _dimension_ops._run_dimension_incremental.

  _incremental_evidence.py → _evidence_parser.py
    Hosts only parse_evidence_from_jsonl. Used by V2's dimension
    runner to build the final Evidence object.

  _incremental_context.py → _analysis_context.py
    Hosts only load_analysis_context. Resolves which dimensions to
    analyze and builds the shared _AnalysisContext.

Updated imports:
  - _loops.py: imports run_dimension_incremental from _dimension_ops
  - _pipeline.py: imports load_analysis_context from _analysis_context
  - cache/dimension_runner.py: imports parse_evidence_from_jsonl
    from _evidence_parser

Pure cosmetic — no behaviour change. Module names now reflect what
they actually contain instead of what they used to participate in.

Full suite: 2757 passing.
